refactor(api): only use Kibana API to sync custom roles
